Testing setup reuses loaded schemas
