+++ lib/polyn/schema_store.rb
@@ -10,10 +10,7 @@ module Polyn
       @nats       = nats
       @store_name = opts[:name] || STORE_NAME
       @key_prefix = "$KV.#{@store_name}"
-      # Accept passed in schemas for test scenarios where we are
-      # doing multiple instantiations and don't need to fetch the
-      # schemas repeatedly
-      @schemas    = opts[:schemas] || fetch_schemas
+      @schemas    = fetch_schemas
     end
 
     attr_reader :schemas
